Fix some details about handling associative array
The responsibility for key value argument in function assoc_insert()
is handed off, so it must be allocated (ie using savestring()) to
prevent inappropriate memory freeing.

Uninitialized associative arrays are invisible and the att_invisible
attribute must be cleared before returning in order for the variable
to be used by the caller.