Pay attention to return of pager_set_margins and respond appropriately.
The program no longer ignores the possibility of irrational margins.
The pager library function pager_set_margins now returns 'false' if
the requested margins exceed the size of the dimension in which they
are declared.  PWB checks the return value and returns a failure code
and sets the error string.