Changes in pwb_action_start() to improve error signalling.
Repaired a logic error that prevented an argument-processing error
from being reported.  While in there, I noticed that other functions
called in the loop never report errors.  I changed the code to embrace
this situation, with notes as to why we're ignoring possible errors
in plotting the screen and handling keypresses.