Refactor action get_dimensions for return variable consistency and optional handle
Calling action get_dimensions without a handle name will return only the
screen width and height.  Providing this allows for displaying status messages
while collecting data, before a pwb handle is available.

The handling of the return variable is changed to conform to other
return variables in the package, using a default variable PWB_ASSOC
if an assoc array is not passed, and using an -A option to set
the variable if the script calls for a named assoc array.