Update prototypes and instructions to recommend using '--' argument.
I forgot how this worked and had trouble with strings leading with '-',
so I updated the documentation to explain now to avoid it.

I _very_ briefly considered updating the code to ignore '-' leading
arguments except for '-v', but quickly reconsidered it as too much
work, having to make custom argument-handling process for measure_string
and limit_print.  Although the '--' argument convention may not be
well-known, it is well-documented and completely solves the problem.