Set left position to 0 for print_left.
In pwbh_position_to_left uses ti_set_cursor_position, which
call tiparm to generate the console-control string to move the
cursor.  tiparm assumes a 0-based index of positions and
converts it to the 1-based coordinates used by the console
code.

The moral is that I should use 0-based coordinates for
ti_set_cursor_position.