Renamed arrays for pwb_read_args, and well as moving pwb get_dimensions after pwb_read_args.
When attempting to show the synopsis with option 'a', there was some
sort of recursion or other failure that prevented any output. Something
was happening during the call to 'pwb get_dimensions' when it came
before pwb_read_args, and stopped happening when I put
'pwb get_dimensions' after.

This fixed the performance, but I don't know why it worked.  It
doesn't make sense.  But at least now you what changed to make it
work, and when.