Update functions pk_get_help_array, pk_prepare_keymap, and pk_get_function.
For these functions, a missing output variable name causes an invisible
error that is difficult to debug.  I added a test at the beginning of these
functions that prints an error message if the output variable is missing.

It would look ugly to an end user, but by helping the developer address
the issue, the end user should never see the message.