Using character DEL/0x7f/control-? for delimiter.
Doing this because none of the other non-printable characters
(except tab, newline, linefeed) are available in XML 1.0, and
runs of tab, newline, and linefeed, despite assurances online,
are treated as a single whitespace character when splitting a
string into array elements.

There is a corresponding change in pwb_txscan that is lost
among many other changes.