Added side margins to center content, progress info in top margin.
I decided that it was too ugly and simple to leave off centering
and a progress indication from the simple page.  Side margins are
pretty easy, especially when broken out into a simple function call.

The side margins broke the simple top margin, so that was redone
to center, with the addition of 'Item x out of x' line to show
progress.