Redone as considerable enhancement of copy of template_divided.
The primary useful part of this template is pwb_keybinder
and execute_line() set of functions, which can be copied
and modified for boilerplate apps.

While the array treatment in here is more extensive than
the keybinder treatment, it is not likely to be useful for
derived apps.  It serves mainly as an example of what can
be done and how.  One important objective is to make the
array stuff easy to prune out in favor of app-specific
data collection and display.