Update set_margins and added global DELIM value.
Improve set margins to always leave at least 2 columns on left
for a scroll bar.

As of now, it seems virtually all data implementations will be
of arrays of joined arrays, and a global DELIM value will be
useful.  Might as well put it in the template to reduce the
differences between templates and their derived scripts.