Changed help key to 'h' like other pagers (less, most, bat).  Fixed help display forwarding.
Help page forwarding is when the user is looking at the help display,
any keypress is recursively forwarded to execution_manager to run
directly.  This was failing because I thought ${@:1} would pass along
all but the first of the function arguments.  Didn't work.  I now use
'shift' and pass all remaining arguments to the same effect.